flask_app/requirements.txt
certifi==2020.6.20
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
SQLAlchemy==1.4.29
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
Flask-Reuploaded==1.2.0
Flask-Uploads==0.2.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
matplotlib==3.5.1
numpy==1.20.0
opencv-python==4.5.5.62
Pillow==9.0.0
